requirements.txt
#upgrade to last version
pytest
Appium-Python-Client
selenium
urllib3
python-dotenv